implementation of get_energy function
based get_energy on the methods in train(), so abstracted shared
components to generateTensorFlowArrays.

In doing this, I realized that the current system implicitly assumed
that there was at least one atom of each type in each batch while
training, which wasn’t great.

I think this is fixed now, by adding a final mask to the results.  if
there isn’t an atom of a given type present in the current image (or
batch of images), it passes a fake fingerprint in, and then sets the
energy to zero of that atom type’s output.

Effectively, unused atom types can be masked.