Put fingerprint range back in.
This goes back to commit a7e4596 in the mainstream Amp project.
At that point an attempt was made to fold fingerprint range
into the initial weights. Going back to right before that point.
Now fingerprint range is part of the model itself.

In testing, this gets rid of the problem that all the input
nodes (first hidden layer) had values of 1.