Bug fix: import matplotlib when class is initiated.
Old method was importing it when the module was first compiled;
new method puts it in the __init__ method of the class so it is
only when the class is initiated.

This was causing a crash with another script that was trying to
start with `matplotlib.use("Agg")`, which was having no effect
because matplotlib had already been opened by the present case.
This caused a much later (and difficult to find) when running in
headless mode.