Easier to catch mistake when missing forces.
Before, if an training image was missing forces, Amp would appear to
hang as the error would occur within a worker image, and the user has to
manually look through every worker's log file to figure out what was
happening.

Now, it pre-checks the images to make sure the expected energies and
forces are there.