I refactored the weights for forces.
I think that it does not make sense to combine weights for energies and
forces because of the way that the kernels are built. I have the strong
opinion that each kernel is its own space and therefore it needs its own
weights. In this sense, I have observed that when this is not the case
then the training becomes almost imposible. Yet to be confirmed.