Refactored sigma's data structure.
If raw user input is a float or list, sigma will be converted into
a dictionary with a "universal" structure for Amp. In the case that
a dict is fed, we check that it complies with how we expect it to be.
This is going to be useful for exploiting anisotropy.

Note that this change should not conflict with previous versions.