Add (possible) random seed to neuralnetwork.
This is turned off by default.
This can be used to make our tests be repeatable, or for benchmarking.
Should solve the problem of the randomly failing numeric-analytic
test which gave slightly different values each time due to different
random initialization of the neural network (making it difficult to
assess where to set the tolerance for agreement).